# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grooming CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
        - name: petId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentFull"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment (supports recurring weeks)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentFull"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentFull"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentFull"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment kanban status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentStatusUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentFull"

  # Financial Entries
...
components:
...
  schemas:
...
    Client:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Pet:
      type: object
      required: [id, clientId, name, size, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        breed:
          type: ["string", "null"]
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    AppointmentFull:
      type: object
      required: [id, tenantId, petId, clientId, scheduledDate, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        petId:
          type: integer
        clientId:
          type: integer
        serviceId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [aguardando, em_atendimento, concluido, cancelado]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        recurringWeeks:
          type: ["integer", "null"]
        recurringGroupId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        pet:
          $ref: "#/components/schemas/Pet"
        client:
          $ref: "#/components/schemas/Client"
        service:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
        package:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
    AppointmentInput:
      type: object
      required: [tenantId, petId, clientId, scheduledDate, totalPrice]
      properties:
        tenantId:
          type: integer
        petId:
          type: integer
        clientId:
          type: integer
        serviceId:
          type: integer
        packageId:
          type: integer
        scheduledDate:
          type: string
          format: date-time
        totalPrice:
          type: number
        notes:
          type: string
        recurringWeeks:
          type: integer
    AppointmentUpdate:
      type: object
      properties:
        petId:
          type: integer
        serviceId:
          type: integer
        packageId:
          type: integer
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [aguardando, em_atendimento, concluido, cancelado]
        totalPrice:
          type: number
        notes:
          type: string
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [aguardando, em_atendimento, concluido, cancelado]

    # Financial Entry
...